requirements.txt
torch
torchvision
timm
numpy
matplotlib
scikit-learn
jupyter
onnx
onnxruntime
gradio
Pillow
seaborn